+++ gxruntime/gxcanvas.cpp
@@ -403,7 +403,7 @@ void gxCanvas::oval( int x1,int y1,int w,int h,bool solid ){
 		y=dest.top-cy;
 		for( int t=dest.top;t<dest.bottom;++y,++t ){
 			float x=sqrt( rsq-y*y )*ar;
-			int xa=floor( cx-x ),xb=floor( cx+x );
+			int xa=(int)floor( cx-x ),xb=(int)floor( cx+x );
 			if( xb<=xa || xa>=viewport.right || xb<=viewport.left ) continue;
 			Rect dr;dr.top=t;dr.bottom=t+1;
 			dr.left=xa<viewport.left ? viewport.left : xa;
@@ -414,14 +414,14 @@ void gxCanvas::oval( int x1,int y1,int w,int h,bool solid ){
 		return;
 	}
 
-	int p_xa,p_xb,t,hh=floor(cy);
+	int p_xa,p_xb,t,hh=(int)floor(cy);
 
-	p_xa=p_xb=cx;
+	p_xa=p_xb=(int)cx;
 	t=dest.top;y=t-cy;
 	if( dest.top>y1 ){ --t;--y; }
 	for( ;t<=hh;++y,++t ){
 		float x=sqrt( rsq-y*y )*ar;
-		int xa=floor( cx-x ),xb=floor( cx+x );
+		int xa=(int)floor( cx-x ),xb=(int)floor( cx+x );
 		Rect r1( xa,t,p_xa-xa,1 );if( r1.right<=r1.left ) r1.right=r1.left+1;
 		if( clip( &r1 ) ) surf->Blt( &r1,0,0,DDBLT_WAIT|DDBLT_COLORFILL,&bltfx );
 		Rect r2( p_xb,t,xb-p_xb,1 );if( r2.left>=r2.right ) r2.left=r2.right-1;
@@ -429,12 +429,12 @@ void gxCanvas::oval( int x1,int y1,int w,int h,bool solid ){
 		p_xa=xa;p_xb=xb;
 	}
 
-	p_xa=p_xb=cx;
+	p_xa=p_xb=(int)cx;
 	t=dest.bottom-1;y=t-cy;
 	if( dest.bottom<y1+h ){ ++t;++y; }
 	for( ;t>hh;--y,--t ){
 		float x=sqrt( rsq-y*y )*ar;
-		int xa=floor( cx-x ),xb=floor( cx+x );
+		int xa=(int)floor( cx-x ),xb=(int)floor( cx+x );
 		Rect r1( xa,t,p_xa-xa,1 );if( r1.right<=r1.left ) r1.right=r1.left+1;
 		if( clip( &r1 ) ) surf->Blt( &r1,0,0,DDBLT_WAIT|DDBLT_COLORFILL,&bltfx );
 		Rect r2( p_xb,t,xb-p_xb,1 );if( r2.left>=r2.right ) r2.left=r2.right-1;
@@ -471,11 +471,11 @@ void gxCanvas::text( int x,int y,const string &t ){
 	if( tx>=viewport.right ) return;
 
 	int b=0,w;
-	while( b<t.size() && tx+(w=font->charWidth( t[b] ))<=viewport.left ){
+	while( b<(int)t.size() && tx+(w=font->charWidth( t[b] ))<=viewport.left ){
 		tx+=w;x+=w;++b;
 	}
 	int e=b;
-	while( e<t.size() && tx<viewport.right ){
+	while( e<(int)t.size() && tx<viewport.right ){
 		tx+=font->charWidth( t[e] );++e;
 	}
